VoiceBot_Ready/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
google-genai==0.2.2
requests==2.31.0
gunicorn==21.2.0